requirements.txt
streamlit==1.41.1
openai==1.59.3
python-dotenv==1.0.1
